Prototype of the bean game
This file is copy the bean game program, and I import some api to let the ball can move, also create some data fields to calculate how many balls are there in the grid.  By the way, I add a space action, when the user press the space then start the game.